[refactor][taier-ui]replace mxGraph import path
